deps: Update CDS library and Node version
The used version of "@sap/cds-dk" was using a non fixed "@sap/cds" version, which was not longer
compatible with the used node version. Because of this these libraries needed to be updated. The
current version of "@sap/cds-dk" is using a fixed version of "@sap/cds", which should prevents this
issue in the future.